pyproject.toml
[project]
name = "powerbi-looker-migration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mariadb>=1.1.14",
    "msal>=1.35.1",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]

[tool.ruff]
target-version = "py313"
line-length = 100
src = [".", "lib", "extraction"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # line length — formatter does not reflow docstrings; avoid churn on long strings
]

[tool.ruff.lint.isort]
known-first-party = ["extraction", "lib"]

# Extraction scripts prepend repo root to sys.path so `lib.*` resolves when run as files; imports follow that bootstrap.
[tool.ruff.lint.per-file-ignores]
"extraction/*.py" = ["E402"]
